+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project that require documentation updates will be d
 ## [NEXT RELEASE]
 - Added a global flag `--token-file` to roxctl causing an API token to be read from the specified file (ROX-2319).
 - Emit warnings if defaults for `--create-updater-sa` or `--collection-method` are used for roxctl (ROX-5264).
+- UI: Omit Cluster column from Deployments list when entity context includes Namespace (ROX-5207)
 
 ## [46.0]
 - Added the following REST APIs: